==> single_cell/07.diff_expression.seurat.Rmd <==
---
title: "Differential expression"
author: "Sergey Naumenko"
date: "`r Sys.Date()`"
output:
    html_document:
        code_folding: hide
        df_print: paged
        highlight: tango
        theme: default
        number_sections: true
        toc: true
        toc_float:
            collapsed: true
            smooth_scroll: false
params:
    variable_features: 3000
    npcs: 30
---


```{r setup, include=FALSE}
library(knitr)
library(tidyverse)
library(readxl)
library(writexl)
library(Seurat)

ggplot2::theme_set(theme_light(base_size = 14))

opts_chunk[["set"]](
    cache = FALSE,
    dev = c("png", "pdf"),
    error = TRUE,
    highlight = TRUE,
    message = FALSE,
    prompt = FALSE,
    tidy = FALSE,
    warning = FALSE)

```

# Overview
- Principal Investigator: 
- Researcher: 
- Experiment: 
- PCs: `r params$npcs`

```{r, rows.print = 25}
de_file <- "tables/de_clusters.xlsx"
seurat <- readRDS("data/seurat.clusters_neuronal.RDS")
de_results <- NULL
    
for (id1 in levels(seurat))
{
    for (id2 in levels(seurat))
    {
        if (id1 < id2 & id1 != id2){
        de_buf <- FindMarkers(seurat, ident.1 = id1, ident.2 = id2) %>% rownames_to_column(var = "gene")
            de_buf$cluster1 <- id1
            de_buf$cluster2 <- id2
            de_buf <- de_buf %>% select(cluster1, cluster2, gene, p_val, avg_logFC, pct.1, pct.2, p_val_adj)
            de_results <- bind_rows(de_results, de_buf)
        }
    }
}

genes <- read_csv("tables/ensembl_w_description.mouse.csv")
de_results <- de_results %>% left_join(genes, by = c("gene" = "ensembl_gene_id")) %>% 
        select(cluster1, cluster2, external_gene_name, gene, p_val, avg_logFC, pct.1, pct.2, p_val_adj,            external_gene_name, description)
    
write_xlsx(list(de_results), de_file)

```

```{r}
de_results <- NULL
    
for (cluster_id in levels(seurat))
{
    print(cluster_id)
    for (condition in c("susceptible", "resilient")){
            de_buf <- NULL
            de_buf <- try(FindMarkers(seurat, ident.1 = condition, 
                          ident.2 = "control",
        	              group.by = "condition",
        	              subset.ident = cluster_id,
                          verbose = FALSE) %>% rownames_to_column(var = "ensembl_gene_id"), 
                          silent = TRUE)
        # new Seurat just returns an empty dataframe
        if(!inherits(de_buf, "try-error") & nrow(de_buf) > 0)
        {
            de_buf$cluster <- paste(cluster_id, collapse=",")
        	de_buf$condition <- condition
        	de_results <- bind_rows(de_results, de_buf)
        }
    }
}
    
genes <- read_csv("tables/ensembl_w_description.mouse.csv")
de_results <- de_results %>% 
              left_join(genes, by = c("ensembl_gene_id" = "ensembl_gene_id")) %>% 
              select(cluster, condition, external_gene_name, ensembl_gene_id, p_val, avg_log2FC, pct.1, pct.2, p_val_adj, description)
    
write_xlsx(list(de_results), "tables/de_susceptible_resilient_vs_control.xlsx")
```


```{r}
de_results <- NULL
for (cluster_id in levels(seurat))
{
    print(cluster_id)
    for (condition in c("resilient")){
            de_buf <- NULL
            de_buf <- try(FindMarkers(seurat, ident.1 = condition, 
                          ident.2 = "susceptible",
        	              group.by = "condition",
        	              subset.ident = cluster_id,
                          verbose = FALSE) %>% rownames_to_column(var = "ensembl_gene_id"), 
                          silent = TRUE)
        if(!inherits(de_buf, "try-error") & nrow(de_buf) > 0)
        {
        	de_buf$cluster <- paste(cluster_id, collapse=",")
        	de_buf$condition <- condition
        	de_results <- bind_rows(de_results, de_buf)
        }
    }
}
    
genes <- read_csv("tables/ensembl_w_description.mouse.csv")
de_results <- de_results %>% 
              left_join(genes, by = c("ensembl_gene_id" = "ensembl_gene_id")) %>% 
              select(cluster, condition, external_gene_name, ensembl_gene_id, p_val, avg_log2FC, pct.1, pct.2, p_val_adj, description)
    
write_xlsx(list(de_results), "tables/de_resilient_vs_susceptible.xlsx")
```

# R session information
```{r}
sessionInfo()
```
